pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "prismq-idea-sources-content-shorts-youtubeshortsource"
version = "1.0.0"
description = "PrismQ module for gathering idea inspirations from YouTube Shorts as part of the AI-powered content generation ecosystem"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "PrismQ", email = "dev@prismq.com"}
]
keywords = ["prismq", "ai", "content-generation", "youtube", "shorts", "idea-sources"]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.100.0",
    "yt-dlp>=2023.10.13",
    "sqlite-utils>=3.35",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
youtube-shorts-source = "src.cli:main"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
